+++ server/src/main/java/org/opensearch/action/admin/cluster/node/stats/NodesStatsRequest.java
@@ -245,7 +245,8 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
         FILE_CACHE_STATS("file_cache"),
         TASK_CANCELLATION("task_cancellation"),
         SEARCH_PIPELINE("search_pipeline"),
-        RESOURCE_USAGE_STATS("resource_usage_stats");
+        RESOURCE_USAGE_STATS("resource_usage_stats"),
+        REPOSITORIES("repositories");
 
         private String metricName;