+++ server/src/main/java/org/opensearch/node/Node.java
@@ -1114,7 +1114,8 @@ public class Node implements Closeable {
                 searchPipelineService,
                 fileCache,
                 taskCancellationMonitoringService,
-                resourceUsageCollectorService
+                resourceUsageCollectorService,
+                repositoryService
             );
 
             final SearchService searchService = newSearchService(