+++ server/src/main/java/org/opensearch/repositories/blobstore/MeteredBlobStoreRepository.java
@@ -34,12 +34,10 @@ package org.opensearch.repositories.blobstore;
 
 import org.opensearch.cluster.metadata.RepositoryMetadata;
 import org.opensearch.cluster.service.ClusterService;
-import org.opensearch.common.UUIDs;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.indices.recovery.RecoverySettings;
 import org.opensearch.repositories.RepositoryInfo;
 import org.opensearch.repositories.RepositoryStatsSnapshot;
-import org.opensearch.threadpool.ThreadPool;
 
 import java.util.Map;
 
@@ -59,14 +57,7 @@ public abstract class MeteredBlobStoreRepository extends BlobStoreRepository {
         Map<String, String> location
     ) {
         super(metadata, namedXContentRegistry, clusterService, recoverySettings);
-        ThreadPool threadPool = clusterService.getClusterApplierService().threadPool();
-        this.repositoryInfo = new RepositoryInfo(
-            UUIDs.randomBase64UUID(),
-            metadata.name(),
-            metadata.type(),
-            location,
-            threadPool.absoluteTimeInMillis()
-        );
+        this.repositoryInfo = new RepositoryInfo(metadata.name(), metadata.type(), location);
     }
 
     @Override
@@ -78,11 +69,6 @@ public abstract class MeteredBlobStoreRepository extends BlobStoreRepository {
     }
 
     public RepositoryStatsSnapshot statsSnapshot() {
-        return new RepositoryStatsSnapshot(repositoryInfo, stats(), RepositoryStatsSnapshot.UNKNOWN_CLUSTER_VERSION, false);
-    }
-
-    public RepositoryStatsSnapshot statsSnapshotForArchival(long clusterVersion) {
-        RepositoryInfo stoppedRepoInfo = repositoryInfo.stopped(threadPool.absoluteTimeInMillis());
-        return new RepositoryStatsSnapshot(stoppedRepoInfo, stats(), clusterVersion, true);
+        return new RepositoryStatsSnapshot(repositoryInfo, stats(), RepositoryStatsSnapshot.UNKNOWN_CLUSTER_VERSION);
     }
 }